@@ -1,6 +1,6 @@
 module github.com/dsh2dsh/expx-cache
 
-go 1.21
+go 1.22
 
 require (
 	github.com/caarlos0/env/v10 v10.0.0
